# Conflicts:
#	app/src/main/java/com/example/bcnet_app/adapter/RecyclerViewAdapter.java

+++ app/src/main/java/com/example/bcnet_app/models/Localitzacio.java
@@ -48,7 +48,7 @@ public class Localitzacio {
     private Float puntuacio;
 
     private ArrayList<Comment> valoracions;
-    private Float puntuacioCOVID;
+    private Integer puntuacioCOVID;
 
     private String semaforUrl;
 
@@ -75,7 +75,6 @@ public class Localitzacio {
     public String getId() {
         return id;
     }
-
     public void setId(String id) {
         this.id = id;
     }
@@ -113,8 +112,8 @@ public class Localitzacio {
         this.direccio = direccio;
     }
 
-    public Float getPuntuacioCOVID() { return puntuacioCOVID; }
-    public void setPuntuacioCOVID(Float puntuacioCOVID) { this.puntuacioCOVID = puntuacioCOVID; }
+    public Integer getPuntuacioCOVID() { return puntuacioCOVID; }
+    public void setPuntuacioCOVID(Integer puntuacioCOVID) { this.puntuacioCOVID = puntuacioCOVID; }
 
 }
